# perf: improve authority state performance