--- a/src/parser.cpp
+++ b/src/parser.cpp
@@ -172,7 +172,8 @@ namespace ada::parser {
           bool password_token_seen{false};
           do {
             std::string_view view = helpers::substring(url_data, input_position);
-            std::string_view authority_view(view.data(), helpers::find_authority_delimiter(url.is_special(), view));
+            size_t location = url.is_special() ? helpers::find_authority_delimiter_special(view) : helpers::find_authority_delimiter(view);
+            std::string_view authority_view(view.data(), location);
             size_t end_of_authority = input_position + authority_view.size();
             // If c is U+0040 (@), then:
             if ((end_of_authority != input_size) && (url_data[end_of_authority] == '@')) {