Thanks to @miguelteixeiraa and @lemire 

Before

```
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations        GHz cycle/byte cycles/url instructions/byte instructions/cycle instructions/ns instructions/url     ns/url      speed  time/byte   time/url      url/s
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
BasicBench_AdaURL_With_Copy       2342 ns         2341 ns       298997    4.81719    13.6646    1003.73            64.271            4.70347         22.6575           4.721k    208.364 345.132M/s  2.89744ns   212.83ns 4.69858M/s
BasicBench_AdaURL_With_Move       2028 ns         2027 ns       346784      5.063    12.5322    920.545           58.1597            4.64083         23.4965         4.27209k    181.818 398.596M/s  2.50881ns  184.283ns 5.42643M/s

```

After

```
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations        GHz cycle/byte cycles/url instructions/byte instructions/cycle instructions/ns instructions/url     ns/url      speed  time/byte   time/url      url/s
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
BasicBench_AdaURL_With_Copy       2351 ns         2351 ns       300044    4.77282    13.7809    1012.27           64.1646            4.65604         22.2225         4.71318k    212.091 343.716M/s  2.90938ns  213.707ns 4.67931M/s
BasicBench_AdaURL_With_Move       2003 ns         2003 ns       349664     5.0045    12.3874    909.909           58.1361            4.69318          23.487         4.27036k    181.818 403.434M/s  2.47872ns  182.073ns  5.4923M/s
```