@lemire 2023-02-24
@anonrig  This is now very good looking.

Can you have a look at what I did in PR https://github.com/ada-url/ada/pull/232 ? It is possible that you might be able to gain a *little* bit of extra performance by mimicking my crazy code. At the very least, you might enjoy it.