@anonrig 2023-02-25
The benchmark results after SWAR routine is like this. I don't think it's worth the change (with SWAR). What do you think? @lemire 

```
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations        GHz cycle/byte cycles/url instructions/byte instructions/cycle instructions/ns instructions/url     ns/url      speed  time/byte   time/url      url/s
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
BasicBench_AdaURL_With_Copy       2340 ns         2340 ns       296349    5.00578    13.9394    1023.91             64.25            4.60925         23.0729         4.71945k    204.545 345.242M/s  2.89652ns  212.762ns 4.70008M/s
BasicBench_AdaURL_With_Move       2028 ns         2027 ns       341307     5.0605     12.526    920.091           58.1448            4.64193         23.4905           4.271k    181.818 398.599M/s  2.50879ns  184.282ns 5.42647M/s
```